move device.js to resources.js
